Create a separate 'monitoring' package for metrics w/friends.
