+++ config-model/src/main/java/com/yahoo/vespa/model/admin/monitoring/DefaultMetricConsumers.java
@@ -1,6 +1,9 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin;
+package com.yahoo.vespa.model.admin.monitoring;
 
+
+import com.yahoo.vespa.model.admin.monitoring.Metric;
+import com.yahoo.vespa.model.admin.monitoring.MetricsConsumer;
 
 import java.util.ArrayList;
 import java.util.Collection;
