+++ config-model/src/main/java/com/yahoo/vespa/model/admin/monitoring/Metric.java
@@ -1,5 +1,5 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin;
+package com.yahoo.vespa.model.admin.monitoring;
 
 import java.util.HashMap;
 import java.util.Map;
