+++ config-model/src/main/java/com/yahoo/vespa/model/admin/monitoring/Yamas.java
@@ -1,8 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.model.admin;
+package com.yahoo.vespa.model.admin.monitoring;
 
 import java.io.Serializable;
-import java.util.Objects;
 
 /**
  * Properties for yamas monitoring service
