+++ config-model/src/main/java/com/yahoo/vespa/model/container/component/StatisticsComponent.java
@@ -1,7 +1,7 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.model.container.component;
 
-import com.yahoo.vespa.model.admin.MonitoringSystem;
+import com.yahoo.vespa.model.admin.monitoring.MonitoringSystem;
 import com.yahoo.container.StatisticsConfig;
 
 /**
